+++ src/services/images.py
@@ -20,10 +20,6 @@ class ImageService:
     def move_svg(destination_file_path, source_file_path):
         class_instance = FilesHelper(destination_file_path=destination_file_path,
                               source_file_path=source_file_path)
-        print(
-            destination_file_path
-        )
-        print(source_file_path)
         class_instance.copy_file()
 
     @staticmethod
