+++ testing/output/robots.txt
@@ -1,3 +1,4 @@
 User-agent: *
 Allow: /
+
 Sitemap: https://microsoft.com/sitemap.xml
